pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "labrag"
version = "0.1.0"
description = "Lab AI Assistant"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
authors = [
    {name = "Cainã Max Couto da Silva", email = "coutodasilva@wisc.edu"},
]
readme = "README.md"
keywords = []
dependencies = [
    "agentic-doc>=0.2.10",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "firecrawl-py>=2.7.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.18",
    "langgraph>=0.5.0",
    "loguru>=0.7.3",
    "openai>=1.82.1",
    "pandas>=2.3.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "requests>=2.32.0",
    "uvicorn>=0.34.2",
]

[project.optional-dependencies]
app = [
    "streamlit>=1.46.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.hatch.metadata]
allow-direct-references = false

[tool.hatch.build]
packages = ["labrag"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest-cov>=6.1.1",
    "pytest>=8.3.4",
    "ruff>=0.9.7",
]
notebook = [
    "ipykernel>=6.29.5",
]

[tool.uv]
default-groups = ["dev", "notebook"]

[tool.ruff]
line-length = 88   # Like Black, use 88 characters per line.
indent-width = 4   # Like Black, use 4 spaces per indentation level.
exclude = ["*.ipynb"] # Exclude Jupyter notebooks from linting.

[tool.ruff.lint]
select = [
    "F",      # Pyflakes
    "E", "W", # pycodestyle
    "I",      # isort
    "N",      # pep8-naming
    "Q",      # flake8-quotes
    "UP",     # pyupgrade
    # "D",      # pydocstyle
    "RUF",    # Ruff-specific rules
    "B",      # flake8-bugbear
    "T20",    # flake8-print
    "C90",    # mccabe (complex structures)
    "SIM",    # flake8-simplify
    "ANN",    # flake8-annotations
    "TID",    # flake8-tidy-imports
]
ignore = [] # ignore specific rules here

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
